+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+## v1.5.1 (2021-08-21)
+### Feature
+* Add sample unit test
+* Add another email sample design
+### Fix
+* Fix wrong root path in constant variable
+
 ## v1.5.0 (2021-07-24)
 ### Feature
 * Add sample unit test